Welcome !
This is the source code of ​​my app <LoveSpeech>. I had the idea around 2006, but was prevented from implementing it. Now, here it is.

With this app you encrypt your shitposts, which are then secured with a password, disguised as foreign characters and decorated with emojis to make it look funnier. Only friends will be able to read the message. For everyone else it will remain a mysterious miracle.

There are 25 different app themes for every user.

There are different styles such as Russian, Arabic, Chinese, Korean, Hieroglyphics, Braille of the Blind, Sumerian, Feng Shui and African Bamum Symbols.

You can export or import password lists. You export your password list, encrypt it and post it somewhere on the Internet. This way others can import them.

Emojis can be turned on or off at will, depending on the mood.